renamed some files, added configuration options
