--- src/tokens.ts
@@ -1,5 +1,0 @@
-import { InjectionToken } from '@angular/core';
-
-export function LocalForageToken() {
-  return new InjectionToken('localforage');
-}
